+++ aristotle_mdr/contrib/autocomplete/views.py
@@ -102,7 +102,7 @@ class UserAutocomplete(autocomplete.Select2QuerySetView):
 
     def get_queryset(self):
         User = get_user_model()
-        
+
         # Don't forget to filter out results depending on the visitor !
         if not self.request.user.is_authenticated():
             raise PermissionDenied