+++ aristotle_mdr/urls/aristotle.py
@@ -78,10 +78,12 @@ urlpatterns=[
     url(r'^workgroup/(?P<iid>\d+)/members/?$', views.workgroups.members, name='workgroupMembers'),
     url(r'^workgroup/(?P<iid>\d+)/items/?$', views.workgroups.items, name='workgroupItems'),
     url(r'^workgroup/(?P<iid>\d+)/leave/?$', views.workgroups.leave, name='workgroup_leave'),
-    url(r'^workgroup/create/?$', views.workgroups.CreateWorkgroup.as_view(), name='workgroup_leave'),
     url(r'^workgroup/addMembers/(?P<iid>\d+)$', views.workgroups.add_members, name='addWorkgroupMembers'),
     url(r'^workgroup/(?P<iid>\d+)/archive/?$', views.workgroups.archive, name='archive_workgroup'),
     url(r'^action/remove/WorkgroupRole/(?P<iid>\d+)/(?P<role>[A-Za-z\-]+)/(?P<userid>\d+)/?$', views.workgroups.remove_role, name='removeWorkgroupRole'),
+
+    url(r'^workgroup/create/?$', views.workgroups.CreateWorkgroup.as_view(), name='workgroup_create'),
+    url(r'^workgroups/all/?$', views.workgroups.ListWorkgroup.as_view(), name='workgroup_list'),
 
     url(r'^discussions/?$', views.discussions.all, name='discussions'),
     url(r'^discussions/new/?$', views.discussions.new, name='discussionsNew'),